- softsigner refactor checkpoint
