+++ ConseilJS-softsigner/README.md
@@ -1,0 +1,1 @@
+# ConseilJS-types
