+++ ConseilJS-softsigner/package.json
@@ -1,0 +1,173 @@
+{
+  "name": "conseiljs-softsigner",
+  "version": "5.0.0-beta.0",
+  "description": "Client-side library for dApp development.",
+  "browser": "dist/index-web.js",
+  "main": "dist/index.js",
+  "types": "dist/index.d.ts",
+  "files": [
+    "dist"
+  ],
+  "scripts": {
+    "test": "mocha \"test/**/*.spec.ts\"",
+    "integration-test": "mocha -r source-map-support/register -r ts-node/register --full-trace -t 600000 -b false \"integration_test/**/*.spec.ts\"",
+    "format": "eslint --fix-dry-run src/*/*.ts",
+    "coverage": "nyc mocha \"test/**/*.spec.ts\" || nyc report | coveralls",
+    "integration-coverage": "nyc mocha -r source-map-support/register -r ts-node/register --full-trace -t 600000 -b false \"integration_test/**/*.spec.ts\" && nyc report | coveralls",
+    "release-check": "npm run test && npm audit && npm outdated",
+    "doc": "rm -rf ./tsdoc && typedoc --readme none --excludeNotExported --mode file --theme markdown --out ./tsdoc ./src",
+    "compile-tezos-parsers": "nearleyc grammar/tezos/Micheline.ne -o src/chain/tezos/lexer/Micheline.ts && nearleyc grammar/tezos/Michelson.ne -o src/chain/tezos/lexer/Michelson.ts && nearleyc grammar/tezos/EntryPointTemplate.ne -o src/chain/tezos/lexer/EntryPointTemplate.ts",
+    "package": "rm -rf ./dist && npm run compile-tezos-parsers && tsc && copyfiles -u 1 \"./src/**/*.js\" dist/",
+    "package-web": "webpack && rm -rf ./dist-web/dist && openssl dgst -sha384 -binary ./dist-web/conseiljs.min.js | openssl base64 -A"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/Cryptonomic/ConseilJS.git"
+  },
+  "keywords": [
+    "tezos",
+    "conseil",
+    "blockchain analytics",
+    "blockchain reporting",
+    "blockchain data access",
+    "web3",
+    "dapp"
+  ],
+  "author": "Cryptonomic Inc",
+  "license": "Apache-2.0",
+  "licenses": [
+    {
+      "type": "Apache-2.0",
+      "url": "http://www.apache.org/licenses/LICENSE-2.0"
+    }
+  ],
+  "bugs": {
+    "url": "https://github.com/Cryptonomic/ConseilJS/issues"
+  },
+  "homepage": "https://cryptonomic.github.io/ConseilJS/#/",
+  "nyc": {
+    "include": [
+      "src/**/*.ts"
+    ],
+    "exclude": [
+      "src/**/lexer/*.*"
+    ],
+    "extension": [
+      ".ts"
+    ],
+    "require": [
+      "ts-node/register"
+    ],
+    "reporter": [
+      "text-lcov"
+    ],
+    "sourceMap": true,
+    "instrument": true
+  },
+  "dependencies": {
+    "big-integer": "1.6.48",
+    "bip32-path": "0.4.2",
+    "bip39": "3.0.2",
+    "blakejs": "1.1.0",
+    "bs58check": "2.1.2",
+    "libsodium-wrappers-sumo": "0.7.6"
+  },
+  "devDependencies": {
+    "@types/chai": "4.2.11",
+    "@types/chai-as-promised": "7.1.2",
+    "@types/mocha": "7.0.2",
+    "@types/nock": "11.1.0",
+    "@types/node": "13.9.1",
+    "@types/node-fetch": "2.5.6",
+    "@types/zxcvbn": "4.4.0",
+    "@typescript-eslint/eslint-plugin": "2.29.0",
+    "@typescript-eslint/parser": "2.29.0",
+    "awesome-typescript-loader": "5.2.1",
+    "chai": "4.2.0",
+    "chai-as-promised": "7.1.1",
+    "copyfiles": "2.2.0",
+    "coveralls": "3.0.12",
+    "eslint": "6.8.0",
+    "eslint-config-airbnb-base": "14.1.0",
+    "eslint-plugin-import": "2.20.2",
+    "glob": "7.1.6",
+    "loglevel": "1.6.8",
+    "mocha": "7.1.1",
+    "nock": "12.0.3",
+    "node-fetch": "2.6.0",
+    "nyc": "15.0.1",
+    "request-promise": "4.2.5",
+    "source-map-loader": "0.2.4",
+    "source-map-support": "0.5.18",
+    "terser-webpack-plugin": "2.3.5",
+    "ts-loader": "6.2.1",
+    "ts-node": "8.9.0",
+    "tsconfig-paths-webpack-plugin": "3.2.0",
+    "typedoc": "0.17.4",
+    "typedoc-plugin-markdown": "2.2.17",
+    "typescript": "3.8.3",
+    "webpack": "4.43.0",
+    "webpack-cli": "3.3.11",
+    "webpack-node-externals": "1.7.2"
+  },
+  "engines": {
+    "node": ">=12.14.x",
+    "npm": ">=6.13.x"
+  },
+  "eslintConfig": {
+    "parser": "@typescript-eslint/parser",
+    "plugins": [
+      "@typescript-eslint"
+    ],
+    "env": {
+      "mocha": true,
+      "node": true
+    },
+    "overrides": [
+      {
+        "files": [
+          "**/**/*.ts"
+        ],
+        "excludedFiles": [
+          "lexer/*.ts"
+        ]
+      }
+    ],
+    "extends": "airbnb-base",
+    "rules": {
+      "comma-dangle": [
+        2,
+        "never"
+      ],
+      "curly": [
+        2,
+        "all"
+      ],
+      "indent": [
+        1,
+        4
+      ],
+      "max-len": [
+        2,
+        {
+          "code": 200,
+          "ignoreTrailingComments": true
+        }
+      ],
+      "no-tabs": [
+        2
+      ],
+      "quotes": [
+        2,
+        "single"
+      ],
+      "array-element-newline": [
+        2,
+        {
+          "multiline": true,
+          "minItems": 5
+        }
+      ]
+    }
+  }
+}
