+++ ConseilJS-softsigner/src/index.ts
@@ -1,0 +1,2 @@
+export * from './FileKeyStore';
+export * from './SoftSigner';
